[BE] fix : 잔 에러 수정
1. Solvedac Service, 3개다 null 일 경우 에러뜨는 현상 수정
2. RequestedCode Entity, Batch에서 검사한 파일인지 확인하는 컬럼 추가
3. MemberInStudyService, 스터디 가입 시 message null로 들어가는 현상 수정

* Solvedac API를 통해 모든 문제를 긁어오려고 할 경우, TooManyRequests 에러 발생...

+++ backend/Aldy/src/main/java/com/example/demo/service/SolvedacServiceImpl.java
@@ -61,9 +61,17 @@ public class SolvedacServiceImpl implements SolvedacService {
 
                     int acceptedUserCount;
                     try {
-                        acceptedUserCount = Integer.parseInt(s.get(8).split(":")[1]);
+                        acceptedUserCount = Integer.parseInt(s.get(6).split(":")[1]);
                     } catch (NumberFormatException e) {
-                        acceptedUserCount = Integer.parseInt(s.get(11).split(":")[1]);
+                        try {
+                            acceptedUserCount = Integer.parseInt(s.get(8).split(":")[1]);
+                        } catch (NumberFormatException ee) {
+                            try {
+                                acceptedUserCount = Integer.parseInt(s.get(11).split(":")[1]);
+                            } catch (NumberFormatException eee) {
+                                acceptedUserCount = Integer.parseInt(s.get(14).split(":")[1]);
+                            }
+                        }
                     }
 
                     problemDtoList.add(new ProblemDto(num, name, acceptedUserCount));
@@ -97,7 +105,7 @@ public class SolvedacServiceImpl implements SolvedacService {
             query.append(")&");
         }
         // "(tag:bfs|tag:dfs)&"
-        if(!algoList.isEmpty()) {
+        if(algoList != null) {
             query.append("(");
             algoList.forEach(a -> {
                 query.append("tag:").append(a).append("|");
@@ -106,7 +114,7 @@ public class SolvedacServiceImpl implements SolvedacService {
             query.append(")&");
         }
         // "!(solved_by:seho27060|solved_by:min61037)"
-        if(!baekjoonIdList.isEmpty()) {
+        if(baekjoonIdList != null) {
             query.append("!(");
             baekjoonIdList.forEach(b -> {
                 query.append("solved_by:").append(b).append("|");
@@ -115,7 +123,9 @@ public class SolvedacServiceImpl implements SolvedacService {
             query.append(")&");
         }
 
-        query.deleteCharAt(query.length() - 1);
+        if(query.length() > 0) {
+            query.deleteCharAt(query.length() - 1);
+        }
 
         return query.toString();
     }
